refactor(a2x): rename SDK package from @a2x/a2x to @a2x/sdk
- rename package for cleaner naming convention
- update all imports and workspace dependencies
- update README with new package name, install instructions, CLI and Client SDK sections
