+++ packages/cli/package.json
@@ -14,7 +14,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@a2x/a2x": "workspace:*",
+    "@a2x/sdk": "workspace:*",
     "commander": "^13.0.0",
     "chalk": "^5.4.0"
   },
