+++ packages/cli/src/commands/a2a/stream.ts
@@ -1,7 +1,7 @@
 import { Command } from 'commander';
 import chalk from 'chalk';
 import crypto from 'node:crypto';
-import type { SendMessageParams } from '@a2x/a2x';
+import type { SendMessageParams } from '@a2x/sdk';
 import { printStatusUpdate, printArtifactChunk, printConnectionError, createClient } from '../../format.js';
 
 export const streamCommand = new Command('stream')
