+++ samples/nextjs/package.json
@@ -9,7 +9,7 @@
     "lint": "eslint"
   },
   "dependencies": {
-    "@a2x/a2x": "workspace:*",
+    "@a2x/sdk": "workspace:*",
     "next": "16.2.3",
     "react": "19.2.4",
     "react-dom": "19.2.4"
